VB script that exports 5 csv files
Working version of VB script, compatible with the user tdurand06 on the lab computers. 
Exports 5 csv files: the field coordinates, the mesh connectivity, the E field, B field and mass density. All csv files are exactly identical to the ones exported using the "Field Extractor" extension of SCMagnet, to the exception that seperate csv files are created, compared to one single csv file for coordinates, connectivity and field values as is the case in SCMagnet.